Move pushpadHelper to helpers
